New version of mkkwd.icn that preserves keyword values.
This version uses the previous contents of keyword.h to preserve
existing keyword values before overwriting the file. New keywords
have a value that is higher than all the existing keywords.
Deleting a keyword leaves a "hole" in the sequence of values.
The definitions are written in alphabetical order, as before.