IB: Prevent em-dashes from interfering with index term matching.
IdxGen allows minuses as part of a word (so "virtual-machine" is
treated as a single word). But a naked em-dash without a space
separating the two words effectively concatenates the words, so in
    Strings in Icon are atomic---there are no operations ...
the 5th word in the sentence is "atomic---there" from idxGen's
point of view, thus defeating a match like
    "(S|s)tring(|s)+5+(A|a)tomic(|ity)".
The solution is to change "atomic---there" into "atomic{---}there".