idxGen: Add \textit to the floating environments and make the float optional.
The previous commit to idxterms.xml solve the problem of makeindex rejecting
entries (so the partial solution of floating them outside environments where
the index hit caused trouble is not necessary). The code has been left in place
and made optional, just in case it comes in handy later on.