Add SNI field to SSL/TLS connections
   SNI (Server Name Indication) is an extension that allows the client
to specify in its ClientHello message the name of the target server.
This is needed when the server hosts several virtual SSL-enabled websites
and needs to identify which certificate they should present to the client,
based on the requested website. Some servers may refuse the connection
if the SNI field is missing even if they host a single website.

Signed-off-by: Jafar Al-Gharaibeh <to.jafar@gmail.com>