Revert Args class back to UniLib behavior
This reverts the behavior of Args to the original behavior in the UniLib version:: program paramters are split into options (start with one or more hyphens) and arguments (don't start with a hyphen).  This makes parsing options easier as well as making it easier to step through non-optiom arguments (e.g. a list of files to process) simpler.