Changes to posix test pty_uni.cin - various
Due to a memeory violation runtime abort, the flush() call was commentted
out. This has allowed the program to run to completion. The runtime flush()
call will need to be looked to see if this is a problem with the pipes
facilities in the iconx runtime.

The output produced from the program has been changed to only show each of
the commands sent and the respective responses received. An additional
procedure has been created to provide a skip and expected response handling.

Signed-off-by: Bruce Rennie <brennie@dcsi.net.au>