Update IPL file numbers.icn to fix large() return value
Previously, a new version of large() was added to handle the versions
of Unicon that had concurrency facilities added. Unlike the the old
version of large(), which is still retained in numbers.icn, the new
version failed to return the supplied parameter. This has been changed
to return the supplied value.

Signed-off-by: Bruce Rennie <brennie@dcsi.net.au>