Update class.icn proc get_class_for()
The procedure get_class_for() will now take a string name as well as an
object value. If the string name is not found in the mapping table, the
procedure will fail.

The string name whether coming from an actual object supplied or from a
string value will be converted to external form first before use as a key
in the mapping table.

Any other kind of value supplied to this procedure will cause failure.

Signed-off-by: Bruce Rennie <brennie@dcsi.net.au>