Update the use code to allow both p__c and p::c reps
Changes to the class.icn, langorocs.icn and predicat.icn to allow the
use of both package__class and package::class forms of the names and have
the same information returned from the respective procedures.

The Class class now stores the external form of a name and will convert the
internal form of a name to the external form before returning the requested
inforamtion. Courtesy of a request and subsequent discussion with Steve
Wampler.

There are possible problems with the UTF8 library but will need to be investigated
later.

Signed-off-by: Bruce Rennie <brennie@dcsi.net.au>