Update equiv.icn to handle more values.
In the case of procedures, files and windows, the original code simply
failed. A change has been made to check if the value produced by the
Unicon function image() indicates that the two procedures, files or
windows are the same in each case. If they are, return the second value,
otherwise fail.

Signed-off-by: Bruce Rennie <brennie@dcsi.net.au>