Add prefix to Unicon functions in other "lang" files.
The other files assocaited with the package "lang" have had the Unicon
functions prefixed with "::" to bring all calls to the same format. It
should be noted that the file property.icn does use the names of some
of the standard Unicon functions as local variables as an example as to
what these changes are preventing.

Signed-off-by: Bruce Rennie <brennie@dcsi.net.au>