install: add resources needed by iconc
rt.a and rt.h shouldn't be in  the bin directory but that is how
we have it now. libtp is also needed but it doesn't get installed.
The only way to get iconc to work now from an installed location is
to invoke iconc directly and pass the location of libtp

iconc -p -I<LIB_HEADER_PATH> -p -L<LIB_PATH>

Signed-off-by: Jafar Al-Gharaibeh <to.jafar@gmail.com>