Update of parser library to standard Unicon productions
This is the update of the parser library to use the statndard Unicon language productions.
The grammar has been modified to incorporate a new semantic action for each production.
The semantic actions are now placed in a separate file unigram.action.icn. The prefix and
suffix code that exitsed in teh original grammar file has now been moved to their respective
code files "unigram.prefix.icn" and "unigram.suffix.icn". This means that the same grammar
file "unigram.y" can be used in both the parser and the Unicon compiler with the specific
semantic actions required for each of the applications being localised to their specific
action, prefix and suffix files.

This shows what can be achieved. The respective changes for the Unicon compiler will be
loaded under another pull request when the testing on that has been completed.

In addition, the parser library class files have been updated to use the new parse tree
that is created by the new format grammar, action, prefix and suffix files.

Signed-off-by: Bruce Rennie <brennie@dcsi.net.au>