Added updated test comparison files for UTF8
The testing for the UTF8 library has been split into 2. The first does not require
operator overloading to be a part of the Unicon build process. The second tests the
operator overloading when present. A change has been made to the uni/lib/Makefile
to ensure that the UTF8 library and support files are compiled during the normal
build process.

Signed-off-by: Bruce Rennie <brennie@dcsi.net.au>