Add direct access methods to the SecureHash plugin.
The methods are additional to the existing direct access procedures.
It is possible to use them in the middle of a sequence of calls to
the Sha method to compute a hash (by violating the class abstraction
to use the context variable), but it probably isn't wise to do so.