Update the tests/general utf8 files
This update replaces the utf8.icn and utf8_ovld.icn files with files that have the
various tests placed in their own specific procedures. This is a workaround to prevent errors
occurring in the macOS run. Further work on the parse tree and various processing elements
in the uniocn parser are to be undertaken to prevent the specific error found from the former
utf8 test files.

Signed-off-by: Bruce Rennie <brennie@dcsi.net.au>