Make auto.h inclusion idempotent.
Custom configuration decisions are usually made via src/h/define.h
But, if the customisation is dependent on the detected configuration,
then access is required to the definitions in auto.h. Unfortunately,
auto.h is normally included __after__ define.h. Making the inclusion
idempotetent means that define.h may include auto.h if it is required.

This also allows define.h to override the automatic configuration if
it is necessary to do so in some circumstances.