Allow experimenting with "parallel" result sequence evaluations
Two new PDCOs that allow experimenting with evaluating result sequences in pseudo-parallel fashion.  In both cases, the result sequences are stepped through and the PDCO suspends once all result sequences have been evaluated to produce a result.  One PDCO fails if any of the sequences is empty and the other fails only after all of the sequences are empty.
Signed-off-by: Steve Wampler <sbw@tapestry.tucson.az.us>