Allow select() to accept a set() as an argument.
select() accepts files or a list of files, but it is useful
to also have support for a set of files. When managing a large
number of sockets it is convenient to maintain the sockets in a set
since a socket can be easily found using member() instead of looping
through a list.

Signed-off-by: Jafar Al-Gharaibeh <to.jafar@gmail.com>