Add classes to uni/lib that implement fixed point arithmetic.
If operator overloading is configured the classes will support it. Otherwise, the
alternative methods (add, sub etc.) can be used in place of the infix operators.
Test code is in tests/unicon. There are two sets of tests, one with overloading and
one without (one is a superset of the other).
Unidoc documentation is embedded in the source code.

Review

No Compiler, No test

squash