Fix ivib to cope with changes in commit 7d4093ae.
7d4093ae changed the compute_absolutes() method in the Dialog class. A side
effect of this change is that the coordinates of ivib's main dialog were no
longer set to (0,0) before the dialog is displayed by custom_show(), which led
to extra space in the dialog window (and an ivib crash under some circumstances).

The smallest, and probably safest, fix is to set the coordinates explicitly in
the custom_show() method.